SDIS_FINAL/requirements.txt
Flask>=3.0.0
pymupdf>=1.24.0
python-docx>=1.1.0
pytesseract>=0.3.10
Pillow>=10.0.0
Werkzeug>=3.0.0
